refactor: move methods related to creating the workspace XML to `RAWebServer.Utilities.WorkspaceBuilder`
